cli: ask before overriding existing files
* Ask user before overriding an existing file using `renku init`
  or `renku runner template`.

(closes #406)

+++ renku/cli/runner.py
@@ -39,17 +39,28 @@ def runner():
 
 
 @runner.command()
+@click.option('--force', is_flag=True, help='Override project templates.')
 @pass_local_client
-def template(client):
+def template(client, force):
     """Render templated configuration files."""
     import pkg_resources
 
     # create the templated files
     for tpl_file in CI_TEMPLATES:
         tpl_path = client.path / tpl_file
-        with tpl_path.open('wb') as dest:
-            with pkg_resources.resource_stream(__name__, tpl_file) as tpl:
-                dest.write(tpl.read())
+        with pkg_resources.resource_stream(__name__, tpl_file) as tpl:
+            content = tpl.read()
+
+            if not force and tpl_path.exists():
+                click.confirm(
+                    'Do you want to override "{tpl_file}"'.format(
+                        tpl_file=tpl_file
+                    ),
+                    abort=True,
+                )
+
+            with tpl_path.open('wb') as dest:
+                dest.write(content)
 
 
 @runner.command()
